+++ crates/stackable-operator/src/cli.rs
@@ -108,6 +108,7 @@
 //!
 use std::{
     ffi::OsStr,
+    ops::Deref,
     path::{Path, PathBuf},
 };
 
@@ -115,10 +116,7 @@ use clap::Args;
 use product_config::ProductConfigManager;
 use snafu::{ResultExt, Snafu};
 
-use crate::{
-    logging::TracingTarget, namespace::WatchNamespace,
-    utils::cluster_info::KubernetesClusterInfoOpts,
-};
+use crate::{namespace::WatchNamespace, utils::cluster_info::KubernetesClusterInfoOpts};
 
 pub const AUTHOR: &str = "Stackable GmbH - info@stackable.tech";
 
@@ -166,6 +164,12 @@ pub enum Command<Run: Args = ProductOperatorRun> {
 ///
 /// ```rust
 /// # use stackable_operator::cli::{Command, ProductOperatorRun, ProductConfigPath};
+/// use clap::Parser;
+/// use stackable_operator::{
+///     cli::TelemetryArguments,
+///     namespace::WatchNamespace,
+/// };
+///
 /// #[derive(clap::Parser, Debug, PartialEq, Eq)]
 /// struct Run {
 ///     #[clap(long)]
@@ -173,11 +177,6 @@ pub enum Command<Run: Args = ProductOperatorRun> {
 ///     #[clap(flatten)]
 ///     common: ProductOperatorRun,
 /// }
-/// use clap::Parser;
-/// use stackable_operator::{
-///     logging::TracingTarget,
-///     namespace::WatchNamespace,
-/// };
 ///
 /// let opts = Command::<Run>::parse_from(["foobar-operator", "run", "--name", "foo", "--product-config", "bar", "--watch-namespace", "foobar"]);
 /// assert_eq!(opts, Command::Run(Run {
@@ -185,7 +184,7 @@ pub enum Command<Run: Args = ProductOperatorRun> {
 ///     common: ProductOperatorRun {
 ///         product_config: ProductConfigPath::from("bar".as_ref()),
 ///         watch_namespace: WatchNamespace::One("foobar".to_string()),
-///         tracing_target: TracingTarget::None,
+///         telemetry_arguments: TelemetryArguments::default(),
 ///         cluster_info_opts: Default::default(),
 ///     },
 /// }));
@@ -195,12 +194,14 @@ pub enum Command<Run: Args = ProductOperatorRun> {
 ///
 /// ```rust
 /// # use stackable_operator::cli::{Command, ProductOperatorRun};
+/// use clap::Parser;
+///
 /// #[derive(clap::Parser, Debug, PartialEq, Eq)]
 /// struct Run {
 ///     #[arg(long)]
 ///     name: String,
 /// }
-/// use clap::Parser;
+///
 /// let opts = Command::<Run>::parse_from(["foobar-operator", "run", "--name", "foo"]);
 /// assert_eq!(opts, Command::Run(Run {
 ///     name: "foo".to_string(),
@@ -217,9 +218,8 @@ pub struct ProductOperatorRun {
     #[arg(long, env, default_value = "")]
     pub watch_namespace: WatchNamespace,
 
-    /// Tracing log collector system
-    #[arg(long, env, default_value_t, value_enum)]
-    pub tracing_target: TracingTarget,
+    #[command(flatten)]
+    pub telemetry_arguments: TelemetryArguments,
 
     #[command(flatten)]
     pub cluster_info_opts: KubernetesClusterInfoOpts,
@@ -245,41 +245,83 @@ impl ProductConfigPath {
     /// Load the [`ProductConfigManager`] from the given path, falling back to the first
     /// path that exists from `default_search_paths` if none is given by the user.
     pub fn load(&self, default_search_paths: &[impl AsRef<Path>]) -> Result<ProductConfigManager> {
-        ProductConfigManager::from_yaml_file(resolve_path(
-            self.path.as_deref(),
-            default_search_paths,
-        )?)
-        .context(ProductConfigLoadSnafu)
+        let resolved_path = Self::resolve_path(self.path.as_deref(), default_search_paths)?;
+        ProductConfigManager::from_yaml_file(resolved_path).context(ProductConfigLoadSnafu)
+    }
+
+    /// Check if the path can be found anywhere
+    ///
+    /// 1. User provides path `user_provided_path` to file. Return [`Error`] if not existing.
+    /// 2. User does not provide path to file -> search in `default_paths` and
+    ///    take the first existing file.
+    /// 3. Return [`Error`] if nothing was found.
+    fn resolve_path<'a>(
+        user_provided_path: Option<&'a Path>,
+        default_paths: &'a [impl AsRef<Path> + 'a],
+    ) -> Result<&'a Path> {
+        // Use override if specified by the user, otherwise search through defaults given
+        let search_paths = if let Some(path) = user_provided_path {
+            vec![path]
+        } else {
+            default_paths.iter().map(|path| path.as_ref()).collect()
+        };
+        for path in &search_paths {
+            if path.exists() {
+                return Ok(path);
+            }
+        }
+        RequiredFileMissingSnafu {
+            search_path: search_paths
+                .into_iter()
+                .map(PathBuf::from)
+                .collect::<Vec<_>>(),
+        }
+        .fail()
     }
 }
 
-/// Check if the path can be found anywhere:
-/// 1) User provides path `user_provided_path` to file -> 'Error' if not existing.
-/// 2) User does not provide path to file -> search in `default_paths` and
-///    take the first existing file.
-/// 3) `Error` if nothing was found.
-fn resolve_path<'a>(
-    user_provided_path: Option<&'a Path>,
-    default_paths: &'a [impl AsRef<Path> + 'a],
-) -> Result<&'a Path> {
-    // Use override if specified by the user, otherwise search through defaults given
-    let search_paths = if let Some(path) = user_provided_path {
-        vec![path]
-    } else {
-        default_paths.iter().map(|path| path.as_ref()).collect()
-    };
-    for path in &search_paths {
-        if path.exists() {
-            return Ok(path);
+#[derive(Debug, Default, PartialEq, Eq, Args)]
+pub struct TelemetryArguments {
+    /// Disable console output.
+    #[arg(long, env)]
+    pub no_console_output: bool,
+
+    /// Enable logging to rolling files located in the specified DIRECTORY.
+    #[arg(long, env, value_name = "DIRECTORY", group = "rolling_logs_group")]
+    pub rolling_logs: Option<PathBuf>,
+
+    /// Time PERIOD after which log files are rolled over.
+    #[arg(long, env, value_name = "PERIOD", requires = "rolling_logs_group")]
+    pub rolling_logs_period: Option<RollingPeriod>,
+
+    /// Enable exporting traces via OTLP.
+    #[arg(long, env)]
+    pub otlp_traces: bool,
+
+    /// Enable exporting logs via OTLP.
+    #[arg(long, env)]
+    pub otlp_logs: bool,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, strum::Display, strum::EnumString, clap::ValueEnum)]
+pub enum RollingPeriod {
+    Minutely,
+    Hourly,
+    Daily,
+    Never,
+}
+
+impl Deref for RollingPeriod {
+    type Target = tracing_appender::rolling::Rotation;
+
+    fn deref(&self) -> &Self::Target {
+        match self {
+            RollingPeriod::Minutely => &tracing_appender::rolling::Rotation::MINUTELY,
+            RollingPeriod::Hourly => &tracing_appender::rolling::Rotation::HOURLY,
+            RollingPeriod::Daily => &tracing_appender::rolling::Rotation::DAILY,
+            RollingPeriod::Never => &tracing_appender::rolling::Rotation::NEVER,
         }
     }
-    RequiredFileMissingSnafu {
-        search_path: search_paths
-            .into_iter()
-            .map(PathBuf::from)
-            .collect::<Vec<_>>(),
-    }
-    .fail()
 }
 
 #[cfg(test)]
@@ -300,6 +342,7 @@ mod tests {
     #[test]
     fn verify_cli() {
         use clap::CommandFactory;
+        ProductOperatorRun::command().print_long_help().unwrap();
         ProductOperatorRun::command().debug_assert()
     }
 
@@ -342,7 +385,7 @@ mod tests {
 
         let file = File::create(full_path_to_create).expect("create temporary file");
 
-        let found_path = resolve_path(
+        let found_path = ProductConfigPath::resolve_path(
             full_user_provided_path.as_deref(),
             &full_default_locations_ref,
         )?;
@@ -358,21 +401,19 @@ mod tests {
     #[test]
     #[should_panic]
     fn resolve_path_user_path_not_existing() {
-        resolve_path(Some(USER_PROVIDED_PATH.as_ref()), &[DEPLOY_FILE_PATH]).unwrap();
+        ProductConfigPath::resolve_path(Some(USER_PROVIDED_PATH.as_ref()), &[DEPLOY_FILE_PATH])
+            .unwrap();
     }
 
     #[test]
     fn resolve_path_nothing_found_errors() {
         if let Err(Error::RequiredFileMissing { search_path }) =
-            resolve_path(None, &[DEPLOY_FILE_PATH, DEFAULT_FILE_PATH])
+            ProductConfigPath::resolve_path(None, &[DEPLOY_FILE_PATH, DEFAULT_FILE_PATH])
         {
-            assert_eq!(
-                search_path,
-                vec![
-                    PathBuf::from(DEPLOY_FILE_PATH),
-                    PathBuf::from(DEFAULT_FILE_PATH)
-                ]
-            )
+            assert_eq!(search_path, vec![
+                PathBuf::from(DEPLOY_FILE_PATH),
+                PathBuf::from(DEFAULT_FILE_PATH)
+            ])
         } else {
             panic!("must return RequiredFileMissing when file was not found")
         }
@@ -381,7 +422,7 @@ mod tests {
     #[test]
     fn product_operator_run_watch_namespace() {
         // clean env var to not interfere if already set
-        env::remove_var(WATCH_NAMESPACE);
+        unsafe { env::remove_var(WATCH_NAMESPACE) };
 
         // cli with namespace
         let opts = ProductOperatorRun::parse_from([
@@ -391,39 +432,30 @@ mod tests {
             "--watch-namespace",
             "foo",
         ]);
-        assert_eq!(
-            opts,
-            ProductOperatorRun {
-                product_config: ProductConfigPath::from("bar".as_ref()),
-                watch_namespace: WatchNamespace::One("foo".to_string()),
-                tracing_target: TracingTarget::None,
-                cluster_info_opts: Default::default(),
-            }
-        );
+        assert_eq!(opts, ProductOperatorRun {
+            product_config: ProductConfigPath::from("bar".as_ref()),
+            watch_namespace: WatchNamespace::One("foo".to_string()),
+            cluster_info_opts: Default::default(),
+            telemetry_arguments: Default::default(),
+        });
 
         // no cli / no env
         let opts = ProductOperatorRun::parse_from(["run", "--product-config", "bar"]);
-        assert_eq!(
-            opts,
-            ProductOperatorRun {
-                product_config: ProductConfigPath::from("bar".as_ref()),
-                watch_namespace: WatchNamespace::All,
-                tracing_target: TracingTarget::None,
-                cluster_info_opts: Default::default(),
-            }
-        );
+        assert_eq!(opts, ProductOperatorRun {
+            product_config: ProductConfigPath::from("bar".as_ref()),
+            watch_namespace: WatchNamespace::All,
+            cluster_info_opts: Default::default(),
+            telemetry_arguments: Default::default(),
+        });
 
         // env with namespace
-        env::set_var(WATCH_NAMESPACE, "foo");
+        unsafe { env::set_var(WATCH_NAMESPACE, "foo") };
         let opts = ProductOperatorRun::parse_from(["run", "--product-config", "bar"]);
-        assert_eq!(
-            opts,
-            ProductOperatorRun {
-                product_config: ProductConfigPath::from("bar".as_ref()),
-                watch_namespace: WatchNamespace::One("foo".to_string()),
-                tracing_target: TracingTarget::None,
-                cluster_info_opts: Default::default(),
-            }
-        );
+        assert_eq!(opts, ProductOperatorRun {
+            product_config: ProductConfigPath::from("bar".as_ref()),
+            watch_namespace: WatchNamespace::One("foo".to_string()),
+            cluster_info_opts: Default::default(),
+            telemetry_arguments: Default::default(),
+        });
     }
 }