+++ crates/stackable-operator/src/product_logging/framework.rs
@@ -1566,12 +1566,9 @@ mod tests {
     #[test]
     fn log4j2_config() {
         let log_config = AutomaticContainerLogConfig {
-            loggers: vec![(
-                "ROOT".to_string(),
-                LoggerConfig {
-                    level: LogLevel::INFO,
-                },
-            )]
+            loggers: vec![("ROOT".to_string(), LoggerConfig {
+                level: LogLevel::INFO,
+            })]
             .into_iter()
             .collect::<BTreeMap<String, LoggerConfig>>(),
             console: Some(AppenderConfig {
@@ -1601,24 +1598,15 @@ mod tests {
     fn log4j2_config_with_additional_loggers() {
         let log_config = AutomaticContainerLogConfig {
             loggers: vec![
-                (
-                    "ROOT".to_string(),
-                    LoggerConfig {
-                        level: LogLevel::INFO,
-                    },
-                ),
-                (
-                    "test".to_string(),
-                    LoggerConfig {
-                        level: LogLevel::INFO,
-                    },
-                ),
-                (
-                    "test_2".to_string(),
-                    LoggerConfig {
-                        level: LogLevel::DEBUG,
-                    },
-                ),
+                ("ROOT".to_string(), LoggerConfig {
+                    level: LogLevel::INFO,
+                }),
+                ("test".to_string(), LoggerConfig {
+                    level: LogLevel::INFO,
+                }),
+                ("test_2".to_string(), LoggerConfig {
+                    level: LogLevel::DEBUG,
+                }),
             ]
             .into_iter()
             .collect::<BTreeMap<String, LoggerConfig>>(),