+++ crates/stackable-webhook/src/tls.rs
@@ -2,25 +2,25 @@
 //! server, which can be used in combination with an Axum [`Router`].
 use std::{net::SocketAddr, sync::Arc};
 
-use axum::{extract::Request, Router};
+use axum::{Router, extract::Request};
 use futures_util::pin_mut;
 use hyper::{body::Incoming, service::service_fn};
 use hyper_util::rt::{TokioExecutor, TokioIo};
 use opentelemetry::trace::{FutureExt, SpanKind};
 use snafu::{ResultExt, Snafu};
-use stackable_certs::{ca::CertificateAuthority, keys::rsa, CertificatePairError};
+use stackable_certs::{CertificatePairError, ca::CertificateAuthority, keys::rsa};
 use stackable_operator::time::Duration;
 use tokio::net::TcpListener;
 use tokio_rustls::{
+    TlsAcceptor,
     rustls::{
+        ServerConfig,
         crypto::aws_lc_rs::default_provider,
         version::{TLS12, TLS13},
-        ServerConfig,
     },
-    TlsAcceptor,
 };
 use tower::{Service, ServiceExt};
-use tracing::{field::Empty, instrument, Instrument, Span};
+use tracing::{Instrument, Span, field::Empty, instrument};
 use tracing_opentelemetry::OpenTelemetrySpanExt;
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
@@ -30,9 +30,7 @@ pub enum Error {
     #[snafu(display("failed to construct TLS server config, bad certificate/key"))]
     InvalidTlsPrivateKey { source: tokio_rustls::rustls::Error },
 
-    #[snafu(display(
-        "failed to create TCP listener by binding to socket address {socket_addr:?}"
-    ))]
+    #[snafu(display("failed to create TCP listener by binding to socket address {socket_addr:?}"))]
     BindTcpListener {
         source: std::io::Error,
         socket_addr: SocketAddr,